Replace more PHPDoc annotations with type hints
Allows us to enable iterable value type checks.

We are now using conditional types for the parameter types in `Options` class.
Unfortunately, PHPStan will currently only use those for checking that correct values are passed arguments – it is not able to infer the arguments in the body so we need to add type coercion comments.